209 - componentize full-screen button
