+++ imports/ui/components/buttons/fullScreenButton.html
@@ -1,0 +1,7 @@
+<template name="fullScreenButton">
+  <button id="fullscreen-button" class="{{className}}" title="{{fullscreen}}">
+    <svg class="player-button-svg" viewBox="0 0 14 14">
+      <path d="M2 9H0v5h5v-2H2V9zM0 5h2V2h3V0H0v5zm12 7H9v2h5V9h-2v3zM9 0v2h3v3h2V0H9z"/>
+    </svg>
+  </button>
+</template>
