+++ imports/ui/pages/player/player.html
@@ -65,11 +65,9 @@
       </div>
     </div>
 
-    <button id="fullscreen-button" class="player-button player-button-fullscreen {{hideControls}}" title="{{fullscreen}}">
-      <svg class="player-button-svg" viewBox="0 0 14 14">
-        <path d="M2 9H0v5h5v-2H2V9zM0 5h2V2h3V0H0v5zm12 7H9v2h5V9h-2v3zM9 0v2h3v3h2V0H9z"/>
-      </svg>
-    </button>
+    <div class="player-button player-button-fullscreen">
+      {{> fullScreenButton className=hideControls target=videoPlayer }}
+    </div>
     {{/unless }}
 
     {{#with video}}
