+++ imports/ui/pages/profile/profile.html
@@ -18,9 +18,7 @@
           <button id="edit-profile">
             {{> icon_gear }}
           </button>
-          <button class="profile-full-screen">
-            {{> icon_full_screen }}
-          </button>
+          {{> fullScreenButton }}
         </div>
       </div>
     </div>
