+++ imports/ui/stylesheets/components/buttons.less
@@ -64,3 +64,7 @@
     }
   }
 }
+
+svg.player-button-svg {
+  fill: @color-grey;
+}
